Review pass 37: completeness, consistency, unambiguity, correctness fixes
Backend:
- GovernanceController.getSpend(): delegate to service to avoid duplicate
  default logic divergence (CFG-070)
- AgentController.getLineage(): replace magic number 10 with LINEAGE_DEPTH_CAP
- DnaProposalService.create(): fix int*long overflow in reviewSla calculation
- OffboardingWalkService.ADMIN_BROADCAST: promote to public constant shared
  with AskService (prevents drift between broadcast targets)
- BackupService: replace RuntimeException-wrapping forEach lambdas with
  explicit helper methods that propagate IOException cleanly
- JwtUtil: add SECONDS_TO_MILLIS constant for seconds-to-ms conversion

Console:
- App.tsx: replace hardcoded 'Single-process mode' with ModeLabel component
  reading VITE_SUMMA_MODE env var
- main.tsx: add routes for board-tasks, triggers, workspaces, nodes,
  role-templates, memory
- api.ts: tighten 13 string-typed enum fields to named union types
- AskInbox.tsx: add escapeHtml() helper; sanitize error, from, to, and
  payload text to prevent XSS from server responses

Docs/Config:
- specs/21-delivery-and-acceptance.md: pin Spring Boot version to 3.4.1
- application-prod.yml: add summa.mode override and show-sql under
  eclipselink properties for parity with base config

+++ backend/src/main/java/com/summa/service/DnaProposalService.java
@@ -40,7 +40,7 @@ public class DnaProposalService {
         if (domainId != null) {
             Optional<DnaDomain> domainOpt = domainService.findById(domainId);
             if (domainOpt.isPresent()) {
-                proposal.setReviewBy(Instant.now().plusSeconds(domainOpt.get().getReviewSlaDays() * 86400L));
+                proposal.setReviewBy(Instant.now().plusSeconds(domainOpt.get().getReviewSlaDays().longValue() * 86400L));
             } else {
                 // Domain not found — use default 7 days
                 proposal.setReviewBy(Instant.now().plusSeconds(DEFAULT_REVIEW_SLA_SECONDS));
